[lineGraph] compiles with clang

+++ src/lineGraph/lineGraph_driver.cpp
@@ -47,9 +47,7 @@ void get_postgres_result(
         std::vector< Line_graph_rt > edge_result,
         Line_graph_rt **return_tuples,
         size_t &sequence) {
-    (*return_tuples) = pgr_alloc(
-            static_cast<int>(edge_result.size()),
-            (*return_tuples));
+    (*return_tuples) = pgr_alloc(edge_result.size(), (*return_tuples));
 
     for (const auto &edge : edge_result) {
         (*return_tuples)[sequence] = {edge.id, edge.source, edge.target,
